API change: Remove train_step_counter arg from `train()`
The train_step_counter is now taken in the TFAgent constructor,
thus avoiding the need to pass it in during the call to `train()`.
This improves compatibility with TF2 as this ensures the
train_step_counter is part of the correct graph.

Note this change required adjusting the TD3Agent loss.

PiperOrigin-RevId: 235799358
Change-Id: I2d7b31a479b0ff8b7ea70eebc0dbe4f7e31c1dfd