Conditional checking of consumed checkpoint in pyTfPolicy restore.
Currently, restoring the pyTfPolicy checks that all variables in the checkpoint match variables in the graph. While this is desired, there could be cases when they don't match - ie. typically the greedy policy is saved, but we may want to restore the exploration policy. The graph for exploration policy may contain local variables not saved in the checkpoint - OUNoise policy is an example. Therefore, it makes sense to enforce the matching only conditionally.

PiperOrigin-RevId: 238006908
Change-Id: I8535e7ab65490dd06355bf4cb589ec0d68589d1a