Added an option for PPO to train on incomplete episodes.
As discussed in #30, right now PPO discards all rewards that were
found on not completed episodes. This commit addresses this issue and
adds an option to train on episodes that were not completed.

Also added tests to ensure that loss is not zero when training on
incomplete episodes depending on the flags.