Stable Tanh bijiector
Because of limited floating point
precision, tanh(9.01) will get 1.0, which cannnot be inversed. A input value
of 9.01 is not the uncommon. Consider a Gaussian distribution with mean=1.0
and std=2. You can get a sample larger than 9.01 or smaller than -9.01 with
probablity 6.2e-5. So we change Tanh.inverse() to make it more stable.