Implemented masked actions in q_network.py and q_policy.py.
This provides support for action masking in environments, where the environment specifies for every observation which actions are valid and which are invalid.

Also reworked target Q-value calculations in DqnAgent and DdqnAgent to use GreedyPolicy instead of tf.reduce_max / tf.argmax. This helps centralize the target value logic in one place and also ensures masked actions are handled correctly.

PiperOrigin-RevId: 263835806
Change-Id: I6068fa810ada0f5a3012ac37d5d47f3a120fb631