Implemented masked actions in categorical_q_network.py and categorical_q_policy.py for the CategoricalDqnAgent (C51).
This provides support for action masking in environments, where the environment specifies for every observation which actions are valid and which are invalid.

Also reworked target Q-value calculations in CategoricalDqnAgent to use GreedyPolicy instead of tf.argmax. This helps centralize the target value logic in one place and also ensures masked actions are handled correctly.

See the previous changes in DqnAgent for more context.

PiperOrigin-RevId: 265777796
Change-Id: I67701aae2b378e6a45f8185765ce2ffe933cac0e