Fix small bug in actor_distribution_network unit test.
Currently the test generates 10 modes and check whether the 9th mode differs from the 10th one, while throwing the other modes away. The "break" only breaks out of the inner loop, and the intention is most likely to break out of the outer loop and make sure when training=True, not all the generated modes are the same.

PiperOrigin-RevId: 293894881
Change-Id: I43171a6d05507b38fac0cdd5994e9aee5e7a76a2