More effective unit tests for behavioral cloning agent by replacing random trajectories.
We were using make_random_trajectory() to generate a trajectory. In graph mode, this emits a Tensor which calls tf.random.uniform further down. Every call of "self.evaluate(replay.run(traj)[0])" runs the learned policy on a brand new random trajectory. Then we compared the output action before and after training, and verified the difference.

This approach does not make sense, as we were running the policy on different trajectories, so the actions would be different even if the agent did not train at all. Here, we modify the test to verify that the loss went down after 10 iterations of training instead.

PiperOrigin-RevId: 293895026
Change-Id: Ie57d94c9c0b170bb323ef930ece26c77e25d94b8