Create PPOKLPenaltyAgent by wrapping the existing PPOAgent.
This agent uses the KL Penalty version of PPO similar to the OpenAI baselines https://github.com/openai/baselines/tree/master/baselines/ppo1

Note that PPOKLPenaltyAgent's behavior can be reproduced by the parent "PPOAgent"
if the right set of parameters are set. However, we strongly encourage future clients to use this agent instead if you rely on the KL penalty version of PPO implementation, because PPOKLPenaltyAgent abstracts away the parameters unrelated to this particular PPO version, making it less prone to errors.

Some parameters were made required in this new agent, when a sensible default value for all environments does not exist. Some defaults (such as reward norm clipping) as the original defaults were optimized for a specific environment.

PiperOrigin-RevId: 297234940
Change-Id: Ied3bc0175e2f13a25fd77e90debb1b2ae1d26e1c