Change PolicySaver.save_checkpoint() to save files in the 'variables' sub-dir.
This makes the layout files for save_checkpoint() the same as for full models,
which is tidier and make it easier to combine will full model files.

Make the code more robust by using constants defined in the saved_model module
for the various components of saved paths. Still use the current values in
tests, because users will likely hardcode these values in their code.

Update all users and tests for save_checkpoint().

PiperOrigin-RevId: 302211945
Change-Id: I45ce03c5feee55e9fe1425f85b789e2216b328f0