[TF-Agents] Add some initial support for non-standard rewards.
The most useful of these is adding a reward_spec arg to time_step_spec() fn.

An important TODO is to not assume single scalar reward tensors in
time_step.{transition, termination, ...}.  Right now we rely on the shape
of the reward input to tell us if we're batched or not; but we should remove
this.

PiperOrigin-RevId: 305042502
Change-Id: I261cede313d45139f84556d817b839c9ab3b0864