Modify distribution seed stream salt in tf_policy.
As _action is not PPO specific, doesn't make sense to give a PPO specific salt.

Also relax the unit tests that incorrectly depended on the seed stream (with a salt) behavior to be deterministic across multiple calls to `.action` when given the same seed, and were coincidentally getting the same actions.

PiperOrigin-RevId: 305328425
Change-Id: Ie9dab376b13088254e7affe38b1be7e877b228a9