[TF-Agents] Replace calls to tf.nest.assert_same_structure with more succinct info.
Prior to this, calls to assert_same_structure would result in really ugly,
long error messages that make it hard to discern the differences in structures.

Now you get an error message that's specific to the piece of code doing the
checking and just prints the two structures:

```
TypeError: action output and policy_step_spec structures do not match:
  PolicyStep(action=., state=(.,.), info=())
vs.
  PolicyStep(action=., state=(), info=())
```
PiperOrigin-RevId: 310926644
Change-Id: Iff7a318e315bf37ec458eddf513eb1fccff9b529