Update PPO Agent to support training experiences with sequence length 1.
Context: currently, with PPO mini batching we feed experiences of shape [B, 2, ...] into training, and drops one of the two steps. This behavior was left from legacy code which calculates advantages during training. However, no data needs to be dropped if advantages are precalculated. This CL makes it so that if advantages were calculated outside of agent.train, no data is wasted.
PiperOrigin-RevId: 313287115
Change-Id: Ic32edd46179b20a8326c49c5597edcdcd6091ee8