[TF-Agents] Add new Keras InnerReshape layer, NestFlatten layer, and NestMap Network.
The InnerReshape layer simplifies sequential combinations of Conv/Dense/RNN
layers, where the required inner dimensions may differ and we want to
be agnostic to the outer rank (e.g., we want it to work if the outer
dimensions are either `[batch_size]`, or `[batch_size, time]`).

The NestMap and NestFlatten networks allows the user to provide a
group of mapper layers, one for each input of an arbitrary spec.
NestMap feeds each input through the associated layer and returns
outputs in the same structure. This can be plugged a Sequence network
with NestFlatten to flatten the outputs of all these layers. The
output of NestFlatten can then be passed to a combiner like
`tf.keras.layers.Add`. NestMap manages any internal RNN/network state
of the individual mappers. These networks form the basis for replacing
the `preprocessing_layers` and `preprocessing_combiner` arguments from
the monolithic TF-Agents Networks.

PiperOrigin-RevId: 319421158
Change-Id: I6b7d59ed52c8aa06c286bc15bfa8aa9b63d4ed4d