Adds gcloud sdk to support nightly testing.
Also changes how LD_LIBRARY_PATH for mujoco is set. On some systems, even with the same version of docker, `ENV LD_LIBRARY_PATH` is not sticky. Research of the issue seems to suggest LD_LIBRARY_PATH is a tricky ENV variable that sometimes is reset. I tested with other ENV variables and they were fine. In my testing, the original Docker worked fine everywhere except webtop. While the issue seems isolated, the approach in this CL is appropriate and works "everywhere".

PiperOrigin-RevId: 339785720
Change-Id: I202530c92a9d4c806ff6d52ad322c6d4e75b1c5e