Add a loss function to TFAgent.
This new method allows the agent to calculate loss without a backwards pass through the networks. This is especially useful when running on TPU with a separate CPU job to gather statistics or save training summaries to TensorBoard.

Include a test util to check that loss() and train() have outputs with the same structure/shape, and optionally same loss value.

PiperOrigin-RevId: 347400955
Change-Id: Ica209e37f45af66b74e03999ff9a0923e4c5dde5