TF-Agents does not require pillow.
gym requires pillow and sets the needed version. Setting this dependency is causing nightly builds to fail. This is caused by gym as of 0.18.0 (released on DEC-18 which is when the releases stopped working) deciding to declare pillow <= 7.2.0 where as before, pillow was not a dependency at all.

https://github.com/openai/gym/commit/eec7d3354a34c0650749923abe12ed0cf28731a3

I do not know why setup.py does not respect the child dependency set by gym but it does not and since TF-Agents does not depend directly on Pillow it is better to just drop it. My check was searching the code base for `pillow` and other that setup.py it is only installed by the colabs which is also not needed now that we have gym as part of tf-agents.

PiperOrigin-RevId: 348505979
Change-Id: I6641daede0b7e8782266cc23cc55ba62cbc0a0ae