[TF Agents] Remove assumptions on format of `training_data_spec`.
The TFAgent data_context construction makes a best-effort attempt to get
the training info_spec.  If it can't get `training_data_spec.policy_info`,
it resolves to an empty tuple.  If we want to use DataConverters with non-
standard training_data_spec in future, we'll have to add a `training_info_spec`
to `TFAgent.__init__`.

PiperOrigin-RevId: 354621416
Change-Id: I8fa812d75c177f4d0446cb8ebc9391b64ac60e2c