Changes eval to use greedy_policy improving half-cheetah-medium-v0 from ~3.2K to ~4.8K average return during 1M step run.
This is still slightly lower than internal numbers which seem to get to 4.9 or 5K and 50-100K faster due to the tf.data shuffle args not being aligned. This also results in the loss curve continuing to get more negative vs. the MLCompass version trending positive after the first 140K or so steps. Another CL will follow that allows the tf.data values to be set or set them via gin.

PiperOrigin-RevId: 395773473
Change-Id: Ic2938a49111e8e165f328ca0a7462cf1f66624ac