Add ability to pass multiple inputs to a single preprocessing layer in EncodingNetwork
Currently, only a single input can be passed to each preprocessing
layer in EncodingNetwork. The current way to get around this when
using custom layers/models as preprocessing layers is to pass a single
tensor and then split it, but this can create large inefficiencies if
the tensors needed are quite different in terms of size/shape. This
patch adds in the ability to use tuples as keys if using a dictionary
of preprocessing layers with the tuple values referencing inputs
in the observation nest.