Changed behavior to allow for flattening of fields
Currently, tensorflow does support flattening of nests with tuples
as keys assuming that all of the keys within the nest are also tuples
so that there aren't any bad comparisons when the sort is performed
internally in Tensorflows tf.nest.flatten function. This patch makes
everything in the preprocessing nest a tuple so that we don't have
to directly set the object with __dict__ and can flatten the fields
so that existing tools like savedmodel will be able to work on these
models.