Fixes LOG_PROBABILITY action info in `RewardPredictionBasePolicy`
The `RewardPredictionBasePolicy` overwrites the `_distribution` method by returning a deterministic distribution of sampled actions, and populating the action probabilities in the returned policy info. However, it sets the `emit_log_probability` parameter in the base TFPolicy class initializer, which causes the base `TFPolicy.action` method to ignore the log probabilities returned by sub-classes, and instead invoke the distribution object returned by sub-classes to obtain the log probabilities. But because `RewardPredictionBasePolicy` returns a deterministic distribution, the log probabilities are always 0.

To fix this, we change `RewardPredictionBasePolicy._distribution` to return the actual distribution instead of a deterministic one based on sampled actions. Because it returns the actual distribution, it can no longer populate `chosen_arm_features` because the actual action sampling has not happened yet. We therefore  also overwrite the `_action` method in `RewardPredictionBasePolicy` to populate `chosen_arm_features` after action sampling.

In addition, we added a few tests for `BoltzmanRewardPredictionPolicy` to verify the expected behavior around its parameters.

PiperOrigin-RevId: 470985015
Change-Id: I0bce336d9ab4b60811401129da87e0d84866d9ac