Supports `max_exploration_probability_hint` in `FalconRewardPredictionPolicy`
This new parameter lets users configure the behavior of the policy by expressing their tolerance on the amount of exploration, which may be more tangible and hence easier to determine than `exploitation_coefficient`. It is an optional float, representing a hint on the maximum exploration probability, internally clipped to [0, 1]. When it is set, `exploitation_coefficient` is ignored and the policy attempts to choose non-greedy actions with at most this probability. When such an upper bound cannot be achieved, e.g. due to insufficient training data, the policy attempts to minimize the probability of choosing non-greedy actions on a best-effort basis.

PiperOrigin-RevId: 516896185
Change-Id: I0a3c27ce23dec426d0293d846fdf305b7caa247b