Register SquashToSpecNormal to an auto-composite tensor.
Previously, this distribution was converted to a composite tensor via the
soon-to-be-removed tfp.experimental.as_composite function. The new composite
tensor format, used via tfp.experimental.auto_composite_tensor, requires an
explicit annotation.

This involved the fulfilling the following requirements on the class that are imposed by AutoCompositeTensor:
1. Making sure the `parameters` field actually corresponds to the initializer
   arguments. See below for questions about this change.
2. Making sure the the initializer arguments have corresponding private instance variables.
3. Adding a `_parameter_properties` to describe the tensor-containing parameters.

This change is *NOT BACKWARDS COMPATIBLE* for two reasons. First, the new
composite tensor on-disk format is not compatible with the old one. Affected users
either need to retrain the model, or do an in-place upgrade which I have ad-hoc
scripts for. Second, the comment on the current implementation implies that the `parameters` intentionaly had `loc` and `scale` placed into it so as to make this wrapper distribution compare equal to the unwrapped distribution. This will no longer work after this change. It's not clear that this was the best way to implement this in the first place, however, since `make_from_parameters` in this file wouldn't have worked on `SquashToSpecNormal` distributions. With this in mind, however, no tests failed after making this change, so perhaps that feature was never used.

PiperOrigin-RevId: 532913627
Change-Id: Icaae615ecea5d192c59680bf07b4dcbad724484c